test: separate getFrequentPlayerNames tests into integration file
- Move 4 failing tRPC endpoint tests to new integration test file
- Fix database initialization pattern using service layer
- Fix tRPC server-side calling pattern using ._def.resolver
- Original test file now has 10 passing tests
- New integration file has 4 passing tests with proper database setup
